backend/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
yfinance>=0.2.30
ta>=0.11.0
gymnasium>=0.28.1
stable-baselines3>=2.1.0
tensorboard
shimmy>=1.1.0
yahoo_fin
fastapi
pydantic
uvicorn
setuptools
streamlit
plotly>=5.18.0
scipy>=1.11.0
scikit-learn>=1.3.0
curl_cffi>=0.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<5
python-multipart>=0.0.6
sqlalchemy>=2.0.0
